Move mockMetadata and mockSwagger to Typescript and place them in __mocks__ directory
