+++ src/__mocks__/mockSwagger.ts
@@ -1,5 +1,4 @@
-// tslint:disable
-
+// tslint:disable:object-literal-sort-keys
 export const mockSwagger = {
   'http://localhost:3000/services/vba_documents/docs/v0/api': {
     openapi: '3.0.0',
